cleaning up logs
