+++ client/components/index.js
@@ -8,6 +8,6 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Signup} from './signup-form'
 export {Login} from './login-form'
-export {default as ChoiceForm} from './choiceForm'
+export {default as ChoiceForm} from './choice-form'
 export {default as UserArticles} from './userArticles'
 export {default as UpdateForm} from './updateForm'
